Enabled reporting all validation errors from Ajv
Users should be able to get all the errors resulting from the validation, not just the first one. Then they can decide how to use the information.

Knowing if a schema is valid or not might be enough for simple use cases, but if we implement this minor modification, we'll be enabling more use cases.